Update vcd_ddl_interrupt_handler.c
Not sure if this is a indicator value or just a info of the level used
but when modified raises the encoder level when
encoding in h264, guys look more at it.

It seems to be highly related with msm_rotator.c
	if (is_rgb)
		perf_level = 2;
	else if (wh <= (640 * 480))
		perf_level = 3;
	else if (wh <= (736 * 1280)) ( 720p or lower use prof4 )
		perf_level = 4;
	else
		perf_level = 5;

This code seems to specify wich level to use, depending on resolutions.

The line on vcd_ddl_interrupt_handler that seems to specify the level ( or the info )
seq_hdr_info.level > VIDC_1080P_H264_LEVEL4) {

To desired level ( remember to add manually higher profiles or wont work ).

[ Profiles ]
-> VIDC_1080P_H264_LEVEL1
-> VIDC_1080P_H264_LEVEL2
-> VIDC_1080P_H264_LEVEL3
-> VIDC_1080P_H264_LEVEL4 >> MAX enabled by default higher missing. ( must add )
-> VIDC_1080P_H264_LEVEL4p1
-> VIDC_1080P_H264_LEVEL4p2
-> VIDC_1080P_H264_LEVEL5
-> VIDC_1080P_H264_LEVEL5p1

It seems to raise the level but bitrate/framerate stills locked up to 25mbps and 30fps
im still working on it.